+++ nucliadb_telemetry/CHANGELOG.rst
@@ -4,7 +4,7 @@ CHANGELOG
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Fix jetstream wrapper
 
 
 1.0.0 (2022-05-16)